+++ src/main/java/real/team12/week5/BasicPlan.java
@@ -1,4 +1,4 @@
-package com.team12_week5;
+package real.team12.week5;
 
 public class BasicPlan {
 	protected double basicFee=0;
