+++ src/main/java/real/team12/week5/Gold.java
@@ -1,4 +1,4 @@
-package com.team12_week5;
+package real.team12.week5;
 
 public class Gold extends BasicPlan{
 	public Gold(double aMinit, double aLine){
